Fix wrong offsetWidth bug
When the pdf-viewer is initialized in a not selected mat-tab (thus is not
visible to the user), "updateSize()" is called with no offsetParent set.
OffsetWidth is 0, so scale will be 1. When switching to the viewer tab
(now viewer is visible) the pdf will have a wrong scaling and will first get
set right, when a change happens.